+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated SysNotify to 1.9.0 and commonized desktop and web notification handling
 - Room list elements now show multiple lines when the font size is large
 - Presence icons for users are now distinguishable by color and shape
+-  Adapted to new mention regex
 
 ### Deprecated